third_party/blink/web_tests/http/tests/xmlhttprequest/xmlhttprequest-mimetype-mixed-case-expected.txt

This tests that XMLHttpRequest properly handles mixed case MIME strings.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.responseXML is non-null.
PASS successfullyParsed is true

TEST COMPLETE
